Initial commit
Created a class in charge of generating fully equipped BST's from an array according to TOP guidelines. In parallel, created mergeSort aux function, as well as some recursive functions to aid in the different methods present inside the tree-generating class. The tree Class possesses insertion and deletion methods, several information methods such as find, findHeight and findDepth to gather data, and also has the ability to print out all its elements in level pre post and in order. Finally, our mother class cancheck if it is balanced, and can rebalance itself.

At the end of the script there is a driver script that puts many of these functionalities to test, as well as some more aux functions to create the conditions for testing.